+++ src/main/java/pizza/product/Product.java
@@ -1,7 +1,5 @@
 package pizza.product;
 
-import java.math.BigDecimal;
-
 public class Product {
 
     //
@@ -12,13 +10,13 @@ public class Product {
 
     private final String name;
 
-    private final BigDecimal price;
+    private final Double price;
 
     //
     // --- constructors ---
     //
 
-    public Product(String productId, String name, BigDecimal price) {
+    public Product(String productId, String name, Double price) {
         this.productId = productId;
         this.name = name;
         this.price = price;
@@ -36,7 +34,7 @@ public class Product {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
