+++ src/main/resources/schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE products
 (
     pk    VARCHAR(10) NOT NULL,
     name  VARCHAR(50) NOT NULL,
-    price DECIMAL(10,2) NOT NULL,
+    price FLOAT       NOT NULL,
     PRIMARY KEY (`pk`)
 );
 
@@ -14,7 +14,7 @@ CREATE TABLE orders
 (
     id          LONG  NOT NULL,
     cst_id      LONG  NOT NULL,
-    total_price DECIMAL(10,2) NOT NULL,
+    total_price FLOAT NOT NULL,
     eta         TIMESTAMP,
     PRIMARY KEY (`id`)
 );
